// src/app/shared/app/models/CustomerService/icustomer-service-policy.ts
export interface CSPolicyData {
	cancelled?: number;
	className?: string;
	clientName?: string;
	clientNo?: number;
	csSpecialConditions?: string;
	insurComp?: string;
	lineOfBusiness?: string;
	periodFrom?: Date;
	periodTo?: Date;
	policiesSNo?: number;
	policyHolder?: string;
	policyNo?: string;
	producer?: string;
	sNo?: number;
}